fix the bug in category page and refactor
