+++ frontend/categories.html
@@ -15,5 +15,4 @@
   </head>
   <body></body>
   <script type="module" src="./js/helpers/Category.js" defer></script>
-  <script type="module" src="./js/view/createNewCategory.js" defer></script>
 </html>
