+++ frontend/login.html
@@ -10,6 +10,6 @@
   <!-- Include jQuery -->
   <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/gsap.min.js"></script>
-  <script src="./js/view/login.js"></script>
+  <script src="./js/helpers/Login.js"></script>
   <body></body>
 </html>
